Moved require old reminders framework decorator.
